requirements.txt
certifi==2017.11.5
chardet==3.0.4
Click==6.7
clickclick==1.2.2
connexion==1.2
Flask==0.12.2
Flask-Cors==3.0.3
py-healthcheck==1.7.0
idna==2.6
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
jsonpickle==0.9.5
jsonschema==2.6.0
MarkupSafe==1.0
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
